# More optimizations on metal