On 30-core M2 Max:

| model                        | backend  | test     |    t/s (Master)|     t/s (PR)   |   Speedup|
|------------------------------|----------|----------|---------------:|---------------:|---------:|
| LLaMA 7B mostly Q4_0         | Metal    | tg 32    |    62.16 ± 0.07| 62.57 ± 0.08   |   1.007  |
| LLaMA 7B mostly Q4_0         | Metal    | tg 64    |    61.65 ± 0.05| 62.09 ± 0.08   |   1.007  |
| LLaMA 7B mostly Q4_0         | Metal    | tg 128   |    61.22 ± 0.13| 61.71 ± 0.05   |   1.008  |
| LLaMA 7B mostly Q4_0         | Metal    | tg 256   |    58.44 ± 0.03| 59.46 ± 0.15   |   1.017  |
| LLaMA 7B mostly Q4_0         | Metal    | pp 32    |   382.65 ± 1.92| 388.64 ± 1.47  |   1.016  |
| LLaMA 7B mostly Q4_0         | Metal    | pp 64    |   450.66 ± 1.26| 468.87 ± 1.91  |   1.040  |
| LLaMA 7B mostly Q4_0         | Metal    | pp 128   |   444.64 ± 0.62| 484.74 ± 0.49  |   1.090  |
| LLaMA 7B mostly Q4_0         | Metal    | pp 256   |   406.50 ± 0.21| 479.20 ± 0.55  |   1.179  |
| LLaMA 7B mostly Q4_0         | Metal    | pp 512   |   327.79 ± 0.18| 433.35 ± 0.17  |   1.322  |
| LLaMA 7B mostly Q4_0         | Metal    | pp 1024  |   227.97 ± 0.10| 352.67 ± 0.04  |   1.547  |

With these changes along with the merged #2951, perplexity now runs in 13.6 minutes on my M2 Max laptop vs ~24 minutes before.