@ggerganov 2023-09-02
M2 Ultra results:

| model         |       size | test       |    master    t/s |    IK        t/s | speedup   |
| --------------| ---------: | ---------- | ---------------: | ---------------: | --------: |
| LLaMA 7B Q4_0 |   3.56 GiB | pp 32      |    701.75 ± 4.36 |    712.61 ± 2.79 |    1.015  |
| LLaMA 7B Q4_0 |   3.56 GiB | pp 64      |    879.19 ± 3.16 |    919.36 ± 5.79 |    1.046  |
| LLaMA 7B Q4_0 |   3.56 GiB | pp 128     |    964.09 ± 1.39 |   1060.88 ± 1.97 |    1.100  |
| LLaMA 7B Q4_0 |   3.56 GiB | pp 256     |    888.98 ± 1.02 |   1073.01 ± 1.69 |    1.207  |
| LLaMA 7B Q4_0 |   3.56 GiB | pp 512     |    719.63 ± 0.23 |   1009.82 ± 0.96 |    1.403  |
| LLaMA 7B Q4_0 |   3.56 GiB | pp 1024    |    549.62 ± 0.86 |    844.49 ± 1.62 |    1.536  |
| LLaMA 7B Q4_0 |   3.56 GiB | tg 128     |     87.45 ± 0.10 |     87.65 ± 0.17 |    1.002  |