@ikawrakow 2023-09-02
With the latest commit https://github.com/ggerganov/llama.cpp/pull/2959/commits/363f0bf5580d9d59aaabfb227f6595f4806b32ab TG for fp16 is basically 2X faster compared to master. PP is also improved by a margin rapidly increasing with context length.

On 30-core M2-Max:

| model                 | backend    | test       |    t/s (Master)  |   t/s (PR)   |   Speedup |
| --------------------- | ---------- | ---------- | ---------------: | -----------: | --------: |
| LLaMA 7B mostly F16   | Metal      | tg 32      |     12.04 ± 0.01 | 24.34 ± 0.05 |  2.022    |
| LLaMA 7B mostly F16   | Metal      | tg 64      |     12.01 ± 0.00 | 24.28 ± 0.03 |  2.022    |
| LLaMA 7B mostly F16   | Metal      | tg 128     |     11.91 ± 0.09 | 24.05 ± 0.11 |  2.020    |
| LLaMA 7B mostly F16   | Metal      | tg 256     |     10.88 ± 0.91 | 23.83 ± 0.11 |  2.190    |
| LLaMA 7B mostly F16   | Metal      | pp 32      |    426.51 ± 2.98 | 433.79 ± 2.54|  1.017    |
| LLaMA 7B mostly F16   | Metal      | pp 64      |    509.79 ± 1.16 | 531.98 ± 1.61|  1.044    |
| LLaMA 7B mostly F16   | Metal      | pp 128     |    503.01 ± 0.99 | 548.36 ± 0.98|  1.090    |
| LLaMA 7B mostly F16   | Metal      | pp 256     |    449.66 ± 0.48 | 543.37 ± 0.49|  1.208    |
| LLaMA 7B mostly F16   | Metal      | pp 512     |    353.80 ± 0.10 | 494.99 ± 0.42|  1.399    |
| LLaMA 7B mostly F16   | Metal      | pp 1024    |    266.73 ± 0.09 | 424.11 ± 0.18|  1.590    |

@ggerganov I'm curious to know how this compares to  #2891 on you M2 Ultra.