@ggerganov 2023-09-02
| model                   |       size | test       |     #2891    t/s |     PR       t/s | speedup   |
| ----------------------- | ---------: | ---------- | ---------------: | ---------------: | --------: |
| LLaMA 7B F16            |  12.55 GiB | pp 32      |    717.82 ± 4.71 |    736.82 ± 2.22 |    1.026  |
| LLaMA 7B F16            |  12.55 GiB | pp 64      |    974.19 ± 3.94 |   1040.40 ± 5.28 |    1.068  |
| LLaMA 7B F16            |  12.55 GiB | pp 128     |   1049.52 ± 1.58 |   1193.71 ± 2.24 |    1.137  |
| LLaMA 7B F16            |  12.55 GiB | pp 256     |    960.61 ± 1.25 |   1214.15 ± 0.90 |    1.264  |
| LLaMA 7B F16            |  12.55 GiB | pp 512     |    763.64 ± 0.41 |   1135.89 ± 0.25 |    1.487  |
| LLaMA 7B F16            |  12.55 GiB | pp 1024    |    574.10 ± 1.05 |    965.57 ± 0.21 |    1.682  |
| LLaMA 7B F16            |  12.55 GiB | tg 32      |     41.57 ± 0.02 |     40.69 ± 0.04 |    0.979  |
| LLaMA 7B F16            |  12.55 GiB | tg 64      |     41.51 ± 0.03 |     40.58 ± 0.03 |    0.978  |
| LLaMA 7B F16            |  12.55 GiB | tg 128     |     41.25 ± 0.03 |     40.27 ± 0.04 |    0.976  |
| LLaMA 7B F16            |  12.55 GiB | tg 256     |     40.88 ± 0.01 |     39.89 ± 0.08 |    0.976  |

Q4_0 7B Perplexity time also dropped - is this expected?

```
perplexity: calculating perplexity over 655 chunks, batch_size=512
perplexity: 0.64 seconds per pass - ETA 6.97 minutes
```

Also, if F16 and Q4_0 `pp` speed is comparable (F16 even faster now), how come the F16 perplexity is so much slower?

```
perplexity: calculating perplexity over 655 chunks, batch_size=512
perplexity: 0.85 seconds per pass - ETA 9.25 minutes
```

Is this a wrong ETA calculation, or do we have some significant overhead somewhere?