@ikawrakow 2023-09-02
> Also, if F16 and Q4_0 pp speed is comparable (F16 even faster now), how come the F16 perplexity is so much slower?

My experience with the ETA is that it is not very accurate. For `Q4_0` it predicted 14.5 minutes but then finished in 13.6 on my M2 Max. For `fp16` it is predicting 15.5 minutes, and I was also wondering why it is slower than `Q4_0` given that t/s for `fp16` is not better than `Q4_0`. I guess, I'll let it run to completion.